Experimental mode added
In the experimental mode every non zero return code causes an
exception NonZeroReturnCodeError to be thrown. User in this mode
is supposed to handle errors not with IF but with EXCEPT.
Rationale for this is that most likely people do not handle errors
for all lines of shell code and there is not way to make them
handle it like in Go. Thus it is easy to skip an error in shell
and continue excecution with it not even noticing it without
verbose output turned on. The mode with exception solves this
problem well. It is a python way and makes you see all the errors.